Use dependencies installed using Nix if possible
All development dependencies should be managed using Nix to ensure
reproducibility. For ease of use, the Nix dependencies are implicitly
used by all scripts.

This may cause a noticeable performance hit if Nix is installed. The
dependencies are assumed to be correctly installed if the environment
variable `IN_MINILUA_NIX_SHELL` is set to a non-empty value (`nix-shell`
exports this automatically). Hence, it's recommended to use the default
Nix workflow of developing inside a `nix-shell` to eliminate this
performance overhead.

Note that `shell.nix` was reformatted using `alejandra` (an
uncompromising Nix formatter) to adhere to standard Nix conventions
(e.g. spaces instead of tabs).